+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.4.4 [unreleased]
 - fix: Set dns resolver as a field, and fix resolved path [PR 103]
+- chore: Remove experimental feature flag
 
 [PR 103]: https://github.com/dariusc93/rust-ipfs/pull/103